+++ js/util.js
@@ -1,0 +1,9 @@
+const getRandomInt = function (first, second) {
+  const min = (first < second) ? Math.ceil(first) : Math.floor(second);
+  const max = (first > second) ? Math.ceil(first) : Math.floor(second);
+  return Math.floor(Math.random() * (max - min + 1)) + min;
+};
+
+const getRandomArrayElement = (elements) => elements[getRandomInt(0, elements.length - 1)];
+
+export { getRandomInt, getRandomArrayElement };
